requirements.txt
pandas>=2.0.0
numpy>=1.24.0
pandas-datareader>=0.10.0
yfinance>=0.2.0
pyyaml>=6.0
click>=8.1.0
flask>=3.0.0
openpyxl>=3.1.0
scipy>=1.11.0
tqdm>=4.65.0
